+++ examples/collatz/src/lib.rs
@@ -16,7 +16,7 @@
 #[cfg(not(test))]
 #[panic_handler]
 fn my_panic(_info: &core::panic::PanicInfo) -> ! {
-    loop {}
+    core::arch::wasm32::unreachable()
 }
 
 #[cfg(not(test))]