+++ tests/integration/src/end_to_end/abi/functions.rs
@@ -25,7 +25,7 @@ fn function_call() {
                 // Required for no-std crates
                 #[panic_handler]
                 fn my_panic(_info: &core::panic::PanicInfo) -> ! {
-                    loop {}
+                    core::arch::wasm32::unreachable()
                 }
 
                 // Required for no-std crates