+++ tests/integration/src/end_to_end/intrinsics/felt.rs
@@ -21,7 +21,7 @@ fn basic_felt_arithmetic() {
                 // Required for no-std crates
                 #[panic_handler]
                 fn my_panic(_info: &core::panic::PanicInfo) -> ! {
-                    loop {}
+                    core::arch::wasm32::unreachable()
                 }
 
                 // Required for no-std crates